# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BattOS API
  version: 0.1.0-alpha
  description: |
    Contract for BattOS Mission Control. System and Memory Core operations are
    implemented today. Product, run and dashboard operations define the v0.1
    contract and are implemented incrementally by roadmap phase.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /status:
    get:
      tags: [System]
      operationId: getStatus
      x-battos-phase: implemented
      x-battos-security: configurable-bearer-implemented
      summary: Command Center status snapshot
      responses:
        "200":
          description: System status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /memory/recent:
    get:
      tags: [Memory]
      operationId: listRecentMemory
      x-battos-phase: implemented
      x-battos-security: configurable-bearer-implemented
      summary: List most recent memory observations
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Recent observations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /memory/search:
    post:
      tags: [Memory]
      operationId: searchMemory
      x-battos-phase: implemented
      x-battos-security: configurable-bearer-implemented
      summary: Full-text memory search with filters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorySearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorySearchResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /memory/save:
    post:
      tags: [Memory]
      operationId: saveMemory
      x-battos-phase: implemented
      x-battos-security: configurable-bearer-implemented
      summary: Save or upsert an observation using topic_key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryObservationInput"
      responses:
        "201":
          description: Observation saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryObservation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /memory/stats:
    get:
      tags: [Memory]
      operationId: getMemoryStats
      x-battos-phase: implemented
      x-battos-security: configurable-bearer-implemented
      summary: Memory aggregate statistics
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /memory/{id}:
    get:
      tags: [Memory]
      operationId: getMemoryObservation
      x-battos-phase: implemented
      x-battos-security: configurable-bearer-implemented
      summary: Get one observation
      parameters:
        - $ref: "#/components/parameters/NumericID"
      responses:
        "200":
          description: Observation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryObservation"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /domains:
    get:
      tags: [Work]
      operationId: listDomains
      x-battos-phase: 3B
      summary: List work domains
      responses:
        "200":
          $ref: "#/components/responses/DomainList"
    post:
      tags: [Work]
      operationId: createDomain
      x-battos-phase: 3B
      summary: Create a domain
      requestBody:
        $ref: "#/components/requestBodies/DomainInput"
      responses:
        "201":
          $ref: "#/components/responses/DomainItem"
  /domains/{id}:
    parameters:
      - $ref: "#/components/parameters/SlugID"
    get:
      tags: [Work]
      operationId: getDomain
      x-battos-phase: 3B
      responses:
        "200":
          $ref: "#/components/responses/DomainItem"
    patch:
      tags: [Work]
      operationId: updateDomain
      x-battos-phase: 3B
      requestBody:
        $ref: "#/components/requestBodies/DomainPatch"
      responses:
        "200":
          $ref: "#/components/responses/DomainItem"
  /projects:
    get:
      tags: [Work]
      operationId: listProjects
      x-battos-phase: 3B
      responses:
        "200":
          $ref: "#/components/responses/ProjectList"
    post:
      tags: [Work]
      operationId: createProject
      x-battos-phase: 3B
      requestBody:
        $ref: "#/components/requestBodies/ProjectInput"
      responses:
        "201":
          $ref: "#/components/responses/ProjectItem"
  /projects/{id}:
    parameters:
      - $ref: "#/components/parameters/SlugID"
    get:
      tags: [Work]
      operationId: getProject
      x-battos-phase: 3B
      responses:
        "200":
          $ref: "#/components/responses/ProjectItem"
    patch:
      tags: [Work]
      operationId: updateProject
      x-battos-phase: 3B
      requestBody:
        $ref: "#/components/requestBodies/ProjectPatch"
      responses:
        "200":
          $ref: "#/components/responses/ProjectItem"
  /goals:
    get:
      tags: [Work]
      operationId: listGoals
      x-battos-phase: 3B
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          $ref: "#/components/responses/GoalList"
    post:
      tags: [Work]
      operationId: createGoal
      x-battos-phase: 3B
      requestBody:
        $ref: "#/components/requestBodies/GoalInput"
      responses:
        "201":
          $ref: "#/components/responses/GoalItem"
  /goals/{id}:
    parameters:
      - $ref: "#/components/parameters/SlugID"
    get:
      tags: [Work]
      operationId: getGoal
      x-battos-phase: 3B
      responses:
        "200":
          $ref: "#/components/responses/GoalItem"
    patch:
      tags: [Work]
      operationId: updateGoal
      x-battos-phase: 3B
      requestBody:
        $ref: "#/components/requestBodies/GoalPatch"
      responses:
        "200":
          $ref: "#/components/responses/GoalItem"
  /tasks:
    get:
      tags: [Work]
      operationId: listTasks
      x-battos-phase: 3B
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          $ref: "#/components/responses/TaskList"
    post:
      tags: [Work]
      operationId: createTask
      x-battos-phase: 3B
      requestBody:
        $ref: "#/components/requestBodies/TaskInput"
      responses:
        "201":
          $ref: "#/components/responses/TaskItem"
  /tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/SlugID"
    get:
      tags: [Work]
      operationId: getTask
      x-battos-phase: 3B
      responses:
        "200":
          $ref: "#/components/responses/TaskItem"
    patch:
      tags: [Work]
      operationId: updateTask
      x-battos-phase: 3B
      requestBody:
        $ref: "#/components/requestBodies/TaskPatch"
      responses:
        "200":
          $ref: "#/components/responses/TaskItem"
  /knowledge/workspaces:
    get:
      tags: [Knowledge]
      operationId: listKnowledgeWorkspaces
      x-battos-phase: 3B
      responses:
        "200":
          description: Knowledge workspaces
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeWorkspaceList"
    post:
      tags: [Knowledge]
      operationId: createKnowledgeWorkspace
      x-battos-phase: 3B
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeWorkspaceInput"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeWorkspace"
  /journals:
    get:
      tags: [Knowledge]
      operationId: listJournals
      x-battos-phase: 3B
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Journals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalList"
    post:
      tags: [Knowledge]
      operationId: createJournal
      x-battos-phase: 3B
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalInput"
      responses:
        "201":
          description: Created journal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Journal"
  /artifacts:
    get:
      tags: [Knowledge]
      operationId: listArtifacts
      x-battos-phase: 3B
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactList"
    post:
      tags: [Knowledge]
      operationId: createArtifact
      x-battos-phase: 3B
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtifactInput"
      responses:
        "201":
          description: Created artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
  /agents:
    get:
      tags: [Agents]
      operationId: listAgents
      x-battos-phase: 3B
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentList"
    post:
      tags: [Agents]
      operationId: createAgent
      x-battos-phase: 3B
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agent-messages:
    post:
      tags: [Messages]
      operationId: sendAgentMessage
      x-battos-phase: B1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAgentMessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMessageEnvelope"
  /agent-messages/{id}/read:
    post:
      tags: [Messages]
      operationId: markAgentMessageRead
      x-battos-phase: B1
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMessageEnvelope"
  /agents/{id}/messages:
    get:
      tags: [Messages]
      operationId: listAgentInbox
      x-battos-phase: B1
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: unread
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Inbox messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMessageList"
  /skills:
    get:
      tags: [Agents]
      operationId: listSkills
      x-battos-phase: 3B
      responses:
        "200":
          description: Skills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillList"
    post:
      tags: [Agents]
      operationId: createSkill
      x-battos-phase: 3B
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
...
  /cli-tools/{id}/install:
    post:
      tags: [Runtimes]
      operationId: requestCliToolInstall
      x-battos-phase: E2
      summary: Request a governed host install for a registered CLI (pending approval)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Install request created awaiting approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliToolInstallEnvelope"
  /cli-tools/{id}/installs:
    get:
      tags: [Runtimes]
      operationId: listCliToolInstalls
      x-battos-phase: E2
      summary: List install requests for a CLI tool, newest first
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Install history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliToolInstallList"
  /cli-tools/installs/{installId}/approve:
    post:
      tags: [Runtimes]
      operationId: approveCliToolInstall
      x-battos-phase: E2
      summary: Decide an install request; approval executes the command on the host
      parameters:
        - name: installId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CliToolInstallDecisionInput"
      responses:
        "200":
          description: Decision recorded (approved installs run in background)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliToolInstallEnvelope"
...
  /repositories:
    get:
      tags: [Repositories]
      operationId: listRepositories
      x-battos-phase: 4C
      responses:
        "200":
          description: Repositories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryList"
    post:
      tags: [Repositories]
      operationId: connectRepository
      x-battos-phase: 4C
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryInput"
      responses:
        "201":
          description: Repository connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
  /runs:
    get:
      tags: [Runs]
      operationId: listRuns
      x-battos-phase: 4B
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
    post:
      tags: [Runs]
      operationId: proposeRun
      x-battos-phase: 4B
      summary: Create a run proposal requiring approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunProposal"
      responses:
        "201":
          description: Proposed run in awaiting_approval state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
  /runs/{id}:
    parameters:
      - $ref: "#/components/parameters/UUID"
    get:
      tags: [Runs]
      operationId: getRun
      x-battos-phase: 4B
      responses:
        "200":
          description: Run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
  /runs/{id}/approvals:
    parameters:
      - $ref: "#/components/parameters/UUID"
    post:
      tags: [Runs]
      operationId: approveRunAction
      x-battos-phase: 4B
      summary: Approve a run, network, commit or push action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Approval recorded and resulting run state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunApprovalResult"
  /runs/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/UUID"
    post:
      tags: [Runs]
      operationId: cancelRun
      x-battos-phase: 4B
      responses:
        "200":
          description: Run cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
  /runs/{id}/logs:
    parameters:
      - $ref: "#/components/parameters/UUID"
    get:
      tags: [Runs]
      operationId: listRunLogs
      x-battos-phase: 4B
      responses:
        "200":
          description: Persisted log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunLogList"
  /runs/{id}/artifacts:
    parameters:
      - $ref: "#/components/parameters/UUID"
    get:
      tags: [Runs]
      operationId: listRunArtifacts
      x-battos-phase: 4B
      responses:
        "200":
          description: Produced artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactList"
  /runs/{id}/diff:
    parameters:
      - $ref: "#/components/parameters/UUID"
    get:
      tags: [Runs]
      operationId: getRunDiff
      x-battos-phase: 4C
      responses:
        "200":
          description: Git diff produced by the run
          content:
            text/plain:
              schema:
                type: string
  /novacore/chat:
    post:
      tags: [NovaCore]
      operationId: chatNovaCore
      x-battos-phase: 5A
      summary: Chat with the optional OS assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Response and proposed actions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: BattOS admin token
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    NumericID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Unauthorized:
      description: Missing or invalid BattOS admin token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    StatusResponse:
      type: object
      required: [version, overall, subsystems, metrics, timestamp]
      properties:
        version:
          $ref: "#/components/schemas/VersionResponse"
        overall:
          $ref: "#/components/schemas/HealthStatus"
        subsystems:
          type: array
          items:
            $ref: "#/components/schemas/SubsystemHealth"
        metrics:
          $ref: "#/components/schemas/SystemMetrics"
        timestamp:
          type: string
          format: date-time
...
    MemorySearchRequest:
      type: object
      required: [query]
      properties:
        query: { type: string }
        filter:
          $ref: "#/components/schemas/MemorySearchFilter"
        limit: { type: integer, minimum: 1, maximum: 200 }
...
    MemoryListResponse:
      type: object
      required: [items, count]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MemoryObservation"
        count: { type: integer }
...